# docs/frontend-backend/openapi-snapshot.yaml
openapi: 3.1.0
info:
  title: BigHead API
  version: 0.3.0
  description: 'Contrato canônico Sprint 3: matriz T01-T56 e implementação FastAPI.'
servers:
- url: http://localhost:8000
  description: Local
paths:
...
  /v1/rooms:
    get:
      tags:
      - collaboration
      summary: Rooms
      operationId: rooms_v1_rooms_get
      security:
      - HTTPBearer: []
      parameters:
      - name: visibility
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Visibility
      - name: cursor
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Cursor
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 50
          title: Limit
      - name: x-organization-id
        in: header
        required: true
        schema:
          type: string
          format: uuid
          title: X-Organization-Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoomListResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '201':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
      x-bighead-screen: T10
      x-bighead-role: member
      x-bighead-cache-event: cursor + counters; `rooms.updated`
      x-bighead-critical-error: sala privada filtrada incorretamente
    post:
      tags:
      - collaboration
      summary: Create Room
      operationId: create_room_v1_rooms_post
      security:
      - HTTPBearer: []
      parameters:
      - name: x-organization-id
        in: header
        required: true
        schema:
          type: string
          format: uuid
          title: X-Organization-Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoomCreateRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Room'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '200':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
      x-bighead-screen: T10
      x-bighead-role: member
      x-bighead-cache-event: cursor + counters; `rooms.updated`
      x-bighead-critical-error: sala privada filtrada incorretamente
  /v1/rooms/{roomId}/messages:
    get:
      tags:
      - collaboration
      summary: Messages
      operationId: messages_v1_rooms__roomId__messages_get
      security:
      - HTTPBearer: []
      parameters:
      - name: roomId
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Roomid
      - name: cursor
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Cursor
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 50
          title: Limit
      - name: x-organization-id
        in: header
        required: true
        schema:
          type: string
          format: uuid
          title: X-Organization-Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageListResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '403':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
      x-bighead-screen: T11
      x-bighead-role: member
      x-bighead-cache-event: realtime + dedupe; `messages.created`
      x-bighead-critical-error: upload/reconnect falho com duplicidade
    post:
      tags:
      - collaboration
      summary: Create Message
      operationId: create_message_v1_rooms__roomId__messages_post
      security:
      - HTTPBearer: []
      parameters:
      - name: roomId
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Roomid
      - name: x-organization-id
        in: header
        required: true
        schema:
          type: string
          format: uuid
          title: X-Organization-Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MessageCreateRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '403':
          description: Room access denied
        '409':
          description: Idempotency conflict
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/rooms/{roomId}/members:
    get:
      tags:
      - collaboration
      summary: Room Members
      operationId: room_members_v1_rooms__roomId__members_get
      security:
      - HTTPBearer: []
      parameters:
      - name: roomId
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Roomid
      - name: x-organization-id
        in: header
        required: true
        schema:
          type: string
          format: uuid
          title: X-Organization-Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoomMemberListResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
      x-bighead-screen: T12
      x-bighead-role: member
      x-bighead-cache-event: leitura sem cache; privacidade da sala por RLS
      x-bighead-critical-error: sala privada invisivel
  /v1/rooms/{roomId}/files:
    get:
      tags:
      - collaboration
      summary: Room Files
      operationId: room_files_v1_rooms__roomId__files_get
      security:
      - HTTPBearer: []
      parameters:
      - name: roomId
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Roomid
      - name: cursor
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Cursor
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 50
          title: Limit
      - name: x-organization-id
        in: header
        required: true
        schema:
          type: string
          format: uuid
          title: X-Organization-Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoomFileListResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '403':
          $ref: '#/components/responses/Problem'
        '410':
          $ref: '#/components/responses/Problem'
      x-bighead-screen: T13
      x-bighead-role: member
      x-bighead-cache-event: signed URL curta; `files.quarantine.updated`
      x-bighead-critical-error: arquivo em quarantine
  /v1/tasks:
    get:
      tags:
      - collaboration
      - tasks
      summary: Tasks
      operationId: tasks_v1_tasks_get
      security:
      - HTTPBearer: []
      parameters:
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/TaskStatus'
          - type: 'null'
          title: Status
      - name: ownerId
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Ownerid
      - name: assigneeId
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Assigneeid
      - name: risk
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/TaskRiskLevel'
          - type: 'null'
          title: Risk
      - name: slaStatus
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/TaskSlaStatus'
          - type: 'null'
          title: Slastatus
      - name: roomId
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Roomid
      - name: cursor
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Cursor
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 50
          title: Limit
      - name: x-organization-id
        in: header
        required: true
        schema:
          type: string
          format: uuid
          title: X-Organization-Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskListResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '403':
          $ref: '#/components/responses/Problem'
      x-bighead-screen: T14
      x-bighead-role: member
      x-bighead-cache-event: cursor; `tasks.updated`; tarefa herda visibilidade da sala
      x-bighead-critical-error: lista vazia sem CTA
    post:
      tags:
      - collaboration
      - tasks
      summary: Create Task
      operationId: create_task_v1_tasks_post
      security:
      - HTTPBearer: []
      parameters:
      - name: Idempotency-Key
        in: header
        required: true
        schema:
          type: string
          minLength: 1
          maxLength: 200
          pattern: .*\S.*
          title: Idempotency-Key
      - name: x-organization-id
        in: header
        required: true
        schema:
          type: string
          format: uuid
          title: X-Organization-Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreateRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskCreateResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '409':
          $ref: '#/components/responses/Problem'
      x-bighead-screen: T15
      x-bighead-role: member
      x-bighead-cache-event: '`Idempotency-Key`; `tasks.created`'
      x-bighead-critical-error: dependencia circular
  /v1/tasks/{taskId}:
    get:
      tags:
      - collaboration
      - tasks
      summary: Task Detail
      operationId: task_detail_v1_tasks__taskId__get
      security:
      - HTTPBearer: []
      parameters:
      - name: taskId
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Taskid
      - name: x-organization-id
        in: header
        required: true
        schema:
          type: string
          format: uuid
          title: X-Organization-Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
      x-bighead-screen: T16
      x-bighead-role: member/reviewer
      x-bighead-cache-event: invalida detalhe; `tasks.transitioned`
      x-bighead-critical-error: conflito `409` preservando comentario
  /v1/runs:
    get:
      tags:
      - collaboration
      - runs
      summary: Runs
      operationId: runs_v1_runs_get
      security:
      - HTTPBearer: []
      parameters:
      - name: taskId
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Taskid
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      - name: cursor
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Cursor
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 50
          title: Limit
      - name: x-organization-id
        in: header
        required: true
        schema:
          type: string
          format: uuid
          title: X-Organization-Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunListResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '403':
          $ref: '#/components/responses/Problem'
      x-bighead-screen: T17
      x-bighead-role: member
      x-bighead-cache-event: polling/realtime; `runs.updated`
      x-bighead-critical-error: lease expirado
  /v1/failures:
    get:
      tags:
      - collaboration
      - runs
      summary: Failures
      operationId: failures_v1_failures_get
      security:
      - HTTPBearer: []
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 50
          title: Limit
      - name: x-organization-id
        in: header
        required: true
        schema:
          type: string
          format: uuid
          title: X-Organization-Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FailureGroupResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '403':
          $ref: '#/components/responses/Problem'
      x-bighead-screen: T18
      x-bighead-role: manager
      x-bighead-cache-event: agregacao por grupo; `failures.updated`
      x-bighead-critical-error: retry indisponivel
  /v1/tasks/calendar:
    get:
      tags:
      - collaboration
      - tasks
      summary: Calendar
      operationId: calendar_v1_tasks_calendar_get
      security:
      - HTTPBearer: []
      parameters:
      - name: from
        in: query
        required: true
        schema:
          type: string
          format: date
          title: From
      - name: to
        in: query
        required: true
        schema:
          type: string
          format: date
          title: To
      - name: ownerIds
        in: query
        required: false
        schema:
          anyOf:
          - type: array
            items:
              type: string
              format: uuid
          - type: 'null'
          title: Ownerids
      - name: x-organization-id
        in: header
        required: true
        schema:
          type: string
          format: uuid
          title: X-Organization-Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskCalendarResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '403':
          $ref: '#/components/responses/Problem'
      x-bighead-screen: T19
      x-bighead-role: manager
      x-bighead-cache-event: calendario; `sla.recomputed`
      x-bighead-critical-error: reagendamento sem permissao
...
  /v1/rooms/{roomId}:
    patch:
      tags:
      - collaboration
      summary: Patch Room
      operationId: patch_room_v1_rooms__roomId__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: roomId
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Roomid
      - name: x-organization-id
        in: header
        required: true
        schema:
          type: string
          format: uuid
          title: X-Organization-Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoomPatchRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoomDetailResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer
  responses:
    Problem:
      description: Erro no formato RFC 9457
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    Room:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        name:
          type: string
          title: Name
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        isPrivate:
          type: boolean
          title: Isprivate
        createdAt:
          type: string
          format: date-time
          title: Createdat
      type: object
      required:
      - id
      - name
      - isPrivate
      - createdAt
      title: Room
    RoomCreateRequest:
      properties:
        name:
          type: string
          maxLength: 160
          minLength: 1
          title: Name
        description:
          anyOf:
          - type: string
            maxLength: 2000
          - type: 'null'
          title: Description
        isPrivate:
          type: boolean
          title: Isprivate
          default: false
      type: object
      required:
      - name
      title: RoomCreateRequest
...
    RoomListResponse:
      properties:
        rooms:
          items:
            $ref: '#/components/schemas/Room'
          type: array
          title: Rooms
        counters:
          additionalProperties:
            type: integer
          type: object
          title: Counters
        nextCursor:
          anyOf:
          - type: string
          - type: 'null'
          title: Nextcursor
      type: object
      required:
      - rooms
      - counters
      title: RoomListResponse
...
    TaskRiskLevel:
      type: string
      enum:
      - low
      - medium
      - high
      - critical
      title: TaskRiskLevel
...